# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/tenants": {
      "get": {
        "operationId": "TenantController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Substring to match against the tenant name or slug.",
            "schema": {
              "maxLength": 100,
              "example": "acme",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum tenants returned.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenants matching the optional name or slug search.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session/JWT."
          },
          "403": {
            "description": "Requires the support-agent role."
          }
        },
        "tags": [
          "tenants"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TenantController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session/JWT."
          }
        },
        "tags": [
          "tenants"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tenants/{id}": {
      "get": {
        "operationId": "TenantController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Legacy/dev tenancy escape when the token has no tenant_id claim. Claim wins when both are present.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant by id (requires matching tenancy context).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session/JWT."
          }
        },
        "tags": [
          "tenants"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "TenantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Legacy/dev tenancy escape when the token has no tenant_id claim. Claim wins when both are present.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant updated after AuthN, tenancy, role, and Permissions AuthZ checks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session/JWT."
          },
          "403": {
            "description": "Requires tenant-admin role and Permissions tuple user:<id>, update, tenant:<id>."
          }
        },
        "tags": [
          "tenants"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me": {
      "get": {
        "operationId": "SingleSignOnController_me",
        "summary": "Current authenticated user (cookie session or Bearer JWT)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "SingleSignOn"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/permissions/check": {
      "post": {
        "operationId": "PermissionsController_check",
        "summary": "Check a fine-grained authorization decision",
        "description": "Evaluates a subject, action, and resource tuple. The stub fails closed until OpenFGA is configured.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckPermissionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session/JWT."
          }
        },
        "tags": [
          "permissions"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/permissions/health": {
      "get": {
        "operationId": "PermissionsController_health",
        "summary": "Report Permissions pillar health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionsHealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "permissions"
        ],
        "security": [
          {
            "oauth2": []
          },
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Platform Kit API",
    "description": "poc-plattform-kit API",
    "version": "0.16.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "x-tenant-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
            "tokenUrl": "/docs/oauth2/token",
            "scopes": {
              "openid": "Sign in",
              "profile": "User profile",
              "offline_access": "Refresh token",
              "platform-kit/.default": "Access Platform Kit API"
            }
          }
        },
        "description": "Microsoft Entra ID (authorization code + PKCE). Register Web redirect URI `/docs/oauth2-redirect.html` on the API app registration. Token exchange is proxied by this API."
      }
    },
    "schemas": {
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        },
        "required": [
          "status"
        ]
      },
      "TenantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "example": "Acme Corp"
          },
          "slug": {
            "type": "string",
            "example": "acme-corp"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Arbitrary tenant settings JSON object, or null."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corp"
          },
          "slug": {
            "type": "string",
            "example": "acme-corp"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corporation"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "CheckPermissionDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "example": "user:alice",
            "description": "OpenFGA subject identifier."
          },
          "action": {
            "type": "string",
            "example": "viewer",
            "description": "OpenFGA relation or permission."
          },
          "resource": {
            "type": "string",
            "example": "document:quarterly-report",
            "description": "OpenFGA resource identifier."
          }
        },
        "required": [
          "subject",
          "action",
          "resource"
        ]
      },
      "CheckPermissionResponseDto": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean",
            "example": false,
            "description": "Whether the subject may perform the action on the resource."
          }
        },
        "required": [
          "allowed"
        ]
      },
      "PermissionsHealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "pillar": {
            "type": "string",
            "example": "permissions"
          }
        },
        "required": [
          "status",
          "pillar"
        ]
      }
    }
  }
}